+++ allauth/socialaccount/tests.py
@@ -1,5 +1,8 @@
+import base64
+import hashlib
 import json
 import random
+import requests
 import warnings
 from urllib.parse import parse_qs, urlparse
 
@@ -166,6 +169,36 @@ class OAuth2TestsMixin(object):
         self.provider = providers.registry.by_id(self.provider_id)
         self.app = setup_app(self.provider)
 
+    def test_provider_has_no_pkce_params(self):
+        provider_settings = app_settings.PROVIDERS.get(self.provider_id, {})
+        provider_settings_with_pkce_set = provider_settings.copy()
+        provider_settings_with_pkce_set["OAUTH_PKCE_ENABLED"] = False
+
+        with self.settings(
+            SOCIALACCOUNT_PROVIDERS={self.provider_id: provider_settings_with_pkce_set}
+        ):
+            self.assertEqual(self.provider.get_pkce_params(), {})
+
+    def test_provider_has_pkce_params(self):
+        provider_settings = app_settings.PROVIDERS.get(self.provider_id, {})
+        provider_settings_with_pkce_set = provider_settings.copy()
+        provider_settings_with_pkce_set["OAUTH_PKCE_ENABLED"] = True
+
+        with self.settings(
+            SOCIALACCOUNT_PROVIDERS={self.provider_id: provider_settings_with_pkce_set}
+        ):
+            pkce_params = self.provider.get_pkce_params()
+            self.assertEqual(
+                set(pkce_params.keys()),
+                {"code_challenge", "code_challenge_method", "code_verifier"},
+            )
+            hashed_verifier = hashlib.sha256(
+                pkce_params["code_verifier"].encode("ascii")
+            )
+            code_challenge = base64.urlsafe_b64encode(hashed_verifier.digest())
+            code_challenge_without_padding = code_challenge.rstrip(b"=")
+            assert pkce_params["code_challenge"] == code_challenge_without_padding
+
     @override_settings(SOCIALACCOUNT_AUTO_SIGNUP=False)
     def test_login(self):
         resp_mock = self.get_mocked_response()
@@ -176,6 +209,50 @@ class OAuth2TestsMixin(object):
             resp_mock,
         )
         self.assertRedirects(resp, reverse("socialaccount_signup"))
+
+    @override_settings(SOCIALACCOUNT_AUTO_SIGNUP=False)
+    def test_login_with_pkce_disabled(self):
+        provider_settings = app_settings.PROVIDERS.get(self.provider_id, {})
+        provider_settings_with_pkce_disabled = provider_settings.copy()
+        provider_settings_with_pkce_disabled["OAUTH_PKCE_ENABLED"] = False
+
+        with self.settings(
+            SOCIALACCOUNT_PROVIDERS={
+                self.provider_id: provider_settings_with_pkce_disabled
+            }
+        ):
+            resp_mock = self.get_mocked_response()
+            if not resp_mock:
+                warnings.warn(
+                    "Cannot test provider %s, no oauth mock" % self.provider.id
+                )
+                return
+            resp = self.login(
+                resp_mock,
+            )
+            self.assertRedirects(resp, reverse("socialaccount_signup"))
+
+    @override_settings(SOCIALACCOUNT_AUTO_SIGNUP=False)
+    def test_login_with_pkce_enabled(self):
+        provider_settings = app_settings.PROVIDERS.get(self.provider_id, {})
+        provider_settings_with_pkce_enabled = provider_settings.copy()
+        provider_settings_with_pkce_enabled["OAUTH_PKCE_ENABLED"] = True
+        with self.settings(
+            SOCIALACCOUNT_PROVIDERS={
+                self.provider_id: provider_settings_with_pkce_enabled
+            }
+        ):
+            resp_mock = self.get_mocked_response()
+            if not resp_mock:
+                warnings.warn(
+                    "Cannot test provider %s, no oauth mock" % self.provider.id
+                )
+                return
+
+            resp = self.login(
+                resp_mock,
+            )
+            self.assertRedirects(resp, reverse("socialaccount_signup"))
 
     def test_account_tokens(self, multiple_login=False):
         if not app_settings.STORE_TOKENS:
@@ -230,13 +307,26 @@ class OAuth2TestsMixin(object):
             + "?"
             + urlencode(dict(process=process))
         )
+
         p = urlparse(resp["location"])
         q = parse_qs(p.query)
+
+        pkce_enabled = app_settings.PROVIDERS.get(self.provider_id, {}).get(
+            "OAUTH_PKCE_ENABLED", False
+        )
+
+        self.assertEqual("code_challenge" in q, pkce_enabled)
+        self.assertEqual("code_challenge_method" in q, pkce_enabled)
+        if pkce_enabled:
+            code_challenge = q["code_challenge"][0]
+            self.assertEqual(q["code_challenge_method"][0], "S256")
+
         complete_url = reverse(self.provider.id + "_callback")
         self.assertGreater(q["redirect_uri"][0].find(complete_url), 0)
         response_json = self.get_login_response_json(
             with_refresh_token=with_refresh_token
         )
+
         if isinstance(resp_mock, list):
             resp_mocks = resp_mock
         else:
@@ -247,6 +337,30 @@ class OAuth2TestsMixin(object):
             *resp_mocks,
         ):
             resp = self.client.get(complete_url, self.get_complete_parameters(q))
+
+            # Find the access token POST request, and assert that it contains
+            # the correct code_verifier if and only if PKCE is enabled
+            request_calls = requests.request.call_args_list
+            for args, kwargs in request_calls:
+                data = kwargs.get("data", {})
+                if (
+                    args[0] == "POST"
+                    and isinstance(data, dict)
+                    and data.get("redirect_uri", "").endswith(complete_url)
+                ):
+                    self.assertEqual("code_verifier" in data, pkce_enabled)
+
+                    if pkce_enabled:
+                        hashed_code_verifier = hashlib.sha256(
+                            data["code_verifier"].encode("ascii")
+                        )
+                        expected_code_challenge = (
+                            base64.urlsafe_b64encode(hashed_code_verifier.digest())
+                            .rstrip(b"=")
+                            .decode()
+                        )
+                        self.assertEqual(code_challenge, expected_code_challenge)
+
         return resp
 
     def get_complete_parameters(self, q):