+++ ChangeLog.rst
@@ -8,6 +8,8 @@ Note worthy changes
 
 - New providers: OpenID Connect.
 
+- Introduced a new provider setting ``OAUTH_PKCE_ENABLED`` that enables the
+  PKCE-enhanced Authorization Code Flow for OAuth 2.0 providers.
 
 0.51.0 (2022-06-07)
 *******************